+++ packages/planet/src/application/planet-application-loader.ts
@@ -3,7 +3,7 @@ import { of, Observable, Subject, forkJoin, from } from 'rxjs';
 import { AssetsLoader } from '../assets-loader';
 import { PlanetApplication, PlanetRouterEvent, SwitchModes, PlanetOptions } from '../planet.class';
 import { switchMap, share, map, tap, distinctUntilChanged, take, filter, catchError } from 'rxjs/operators';
-import { getHTMLElement, coerceArray, createElementByTemplate, debug } from '../helpers';
+import { getHTMLElement, coerceArray, createElementByTemplate } from '../helpers';
 import { PlanetApplicationRef } from './planet-application-ref';
 import { PlanetPortalApplication } from './portal-application';
 import { PlanetApplicationService } from './planet-application.service';
@@ -224,12 +224,12 @@ export class PlanetApplicationLoader {
                                 setTimeout(() => {
                                     // 此处判断是因为如果静态资源加载完毕还未启动被取消，还是会启动之前的应用，虽然可能性比较小，但是无法排除这种可能性，所以只有当 Event 是最后一个才会启动
                                     if (this.startRouteChangeEvent === event) {
-                                        this.ngZone.runOutsideAngular(() => {
-                                            forkJoin(apps$).subscribe(() => {
-                                                this.setLoadingDone();
-                                                this.ensurePreloadApps(apps);
-                                            });
+                                        // this.ngZone.runOutsideAngular(() => {
+                                        forkJoin(apps$).subscribe(() => {
+                                            this.setLoadingDone();
+                                            this.ensurePreloadApps(apps);
                                         });
+                                        // });
                                     }
                                 });
                             } else {
@@ -305,7 +305,6 @@ export class PlanetApplicationLoader {
         app: PlanetApplication,
         defaultStatus: 'hidden' | 'display' = 'display'
     ): Observable<PlanetApplicationRef> {
-        debug(`${app.name} start bootstrap`);
         this.setAppStatus(app, ApplicationStatus.bootstrapping);
         const appRef = getPlanetApplicationRef(app.name);
         if (appRef && appRef.bootstrap) {
@@ -333,7 +332,6 @@ export class PlanetApplicationLoader {
             }
             return result.pipe(
                 tap(() => {
-                    debug(`${app.name} bootstrapped`);
                     this.setAppStatus(app, ApplicationStatus.bootstrapped);
                     if (defaultStatus === 'display' && appRootElement) {
                         appRootElement.removeAttribute('style');
@@ -437,7 +435,6 @@ export class PlanetApplicationLoader {
     preload(app: PlanetApplication, directBootstrap?: boolean): Observable<PlanetApplicationRef> {
         const status = this.appsStatus.get(app);
         if (!status || status === ApplicationStatus.loadError) {
-            debug(`${app.name} start preloading`);
             return this.startLoadAppAssets(app).pipe(
                 switchMap(() => {
                     if (directBootstrap) {
@@ -446,7 +443,6 @@ export class PlanetApplicationLoader {
                         return this.ngZone.runOutsideAngular(() => {
                             return this.takeOneStable().pipe(
                                 switchMap(() => {
-                                    debug(`${app.name} start bootstrap on stable`);
                                     return this.bootstrapApp(app, 'hidden');
                                 })
                             );
@@ -468,12 +464,10 @@ export class PlanetApplicationLoader {
                 }),
                 take(1),
                 map(() => {
-                    debug(`${app.name} assets is loading or app is bootstrapped, return appRef`);
                     return getPlanetApplicationRef(app.name);
                 })
             );
         } else {
-            debug(`${app.name} status is '${status}' , return appRef`);
             return of(getPlanetApplicationRef(app.name));
         }
     }